Move utils.managers -> utils.loaders
